Update for 2014 changes to Ordnances
Grace 2 of 5 February 2014 changed the rule for when General Admission happens,
bring it into line with the dates that had been published in the
acompanying table since 2010. Since the table seems to be being treated
as cannonical, this update reworks the general_admission() method to use
transcribed dates and provides a new general_admission_alg() method that
uses the published algorythm.

This creates version 3.00 of the module.

At the same time make_term_dates_callendar was updatd not to omit General
Admission and the Long vac term for 2025 which were the dates affected
by the inconsitency.